Rename use-cases/recipes to use-cases/recipe.
