+++ gb-server/src/use-cases/recipe/index.js
@@ -1,7 +1,7 @@
 import { BadRequestError, NotFoundError, ConflictError } from "../../entities/helpers/custom-errors";
 import { listAllRecipes, findRecById, deleteRecById } from "../../data-access/public/recipes-db";
 import makeListRecipes from "./list-recipes";
-import makeDeleteOneRecipe from "./delete-recipes";
+import makeDeleteOneRecipe from "./delete-recipe";
 
 const listRecipes = makeListRecipes({ listAllRecipes, findRecById, NotFoundError });
 const deleteOneRecipe = makeDeleteOneRecipe({ findRecById, deleteRecById, BadRequestError, NotFoundError});
